# src/types/handlers/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

export type Course = {
    id: number;
    name: string;
    creatorId: number;
    description: string;
    createdAt: string;
};

export type CourseCreate = {
    name: string;
    creatorId: number;
    description: string;
};

export type Lesson = {
    id: number;
    name: string;
    courseId: number;
    body: string;
    createdAt: string;
};

export type LessonCreate = {
    name: string;
    courseId: number;
    body: string;
};

export type User = {
    id: number;
    fullName: string | null;
    email: string;
    updatedAt: string | null;
    createdAt: string;
};

export type UserCreate = {
    fullName: string | null;
    email: string;
};

export type CoursesIndexData = {
    body?: never;
    path?: never;
    query?: {
        page?: number;
    };
    url: '/courses';
};

export type CoursesIndexResponses = {
    /**
     * The request has succeeded.
     */
    200: {
        data: Array<Course>;
    };
};

export type CoursesIndexResponse = CoursesIndexResponses[keyof CoursesIndexResponses];

export type CoursesCreateData = {
    body: CourseCreate;
    path?: never;
    query?: never;
    url: '/courses';
};

export type CoursesCreateResponses = {
    /**
     * The request has succeeded and a new resource has been created as a result.
     */
    201: Course;
};

export type CoursesCreateResponse = CoursesCreateResponses[keyof CoursesCreateResponses];

export type CoursesDeleteData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/courses/{id}';
};

export type CoursesDeleteResponses = {
    /**
     * There is no content to send for this request, but the headers may be useful.
     */
    204: void;
};

export type CoursesDeleteResponse = CoursesDeleteResponses[keyof CoursesDeleteResponses];

export type CoursesShowData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/courses/{id}';
};

export type CoursesShowResponses = {
    /**
     * The request has succeeded.
     */
    200: Course;
};

export type CoursesShowResponse = CoursesShowResponses[keyof CoursesShowResponses];

export type LessonsIndexData = {
    body?: never;
    path?: never;
    query?: {
        page?: number;
    };
    url: '/lessons';
};

export type LessonsIndexResponses = {
    /**
     * The request has succeeded.
     */
    200: {
        data: Array<Lesson>;
    };
};

export type LessonsIndexResponse = LessonsIndexResponses[keyof LessonsIndexResponses];

export type LessonsCreateData = {
    body: LessonCreate;
    path?: never;
    query?: never;
    url: '/lessons';
};

export type LessonsCreateResponses = {
    /**
     * The request has succeeded and a new resource has been created as a result.
     */
    201: Lesson;
};

export type LessonsCreateResponse = LessonsCreateResponses[keyof LessonsCreateResponses];

export type LessonsDeleteData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/lessons/{id}';
};

export type LessonsDeleteResponses = {
    /**
     * There is no content to send for this request, but the headers may be useful.
     */
    204: void;
};

export type LessonsDeleteResponse = LessonsDeleteResponses[keyof LessonsDeleteResponses];

export type LessonsShowData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/lessons/{id}';
};

export type LessonsShowResponses = {
    /**
     * The request has succeeded.
     */
    200: Lesson;
};

export type LessonsShowResponse = LessonsShowResponses[keyof LessonsShowResponses];

export type UsersIndexData = {
    body?: never;
    path?: never;
    query?: {
        page?: number;
    };
    url: '/users';
};

export type UsersIndexResponses = {
    /**
     * The request has succeeded.
     */
    200: {
        data: Array<User>;
    };
};

export type UsersIndexResponse = UsersIndexResponses[keyof UsersIndexResponses];

export type UsersCreateData = {
    body: UserCreate;
    path?: never;
    query?: never;
    url: '/users';
};

export type UsersCreateResponses = {
    /**
     * The request has succeeded and a new resource has been created as a result.
     */
    201: User;
};

export type UsersCreateResponse = UsersCreateResponses[keyof UsersCreateResponses];

export type UsersDeleteData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/users/{id}';
};

export type UsersDeleteResponses = {
    /**
     * There is no content to send for this request, but the headers may be useful.
     */
    204: void;
};

export type UsersDeleteResponse = UsersDeleteResponses[keyof UsersDeleteResponses];

export type UsersShowData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/users/{id}';
};

export type UsersShowResponses = {
    /**
     * The request has succeeded.
     */
    200: User;
};

export type UsersShowResponse = UsersShowResponses[keyof UsersShowResponses];